Fundamental Kademlia interfaces (#70)
Pull request to discuss the fundamental Kademlia interfaces.
